implemented most of the frontend stuff + routing

+++ src/main/kotlin/com/fhtw/bergziege/article/ArticleRepository.kt
@@ -2,5 +2,5 @@ package com.fhtw.bergziege.article
 
 import org.springframework.data.repository.CrudRepository
 
-public interface ArticleRepository : CrudRepository<Article, Integer> {
+public interface ArticleRepository : CrudRepository<Article, Long> {
 }
