+++ src/main/kotlin/com/fhtw/bergziege/author/AuthorRepository.kt
@@ -3,6 +3,6 @@ package com.fhtw.bergziege.author
 import org.springframework.data.repository.CrudRepository
 
 // auto implemented by Spring, turns into a "Bean" called authorRepository
-public interface AuthorRepository : CrudRepository<Author, Integer> {
+public interface AuthorRepository : CrudRepository<Author, Long> {
     public fun findByName(name: String): List<Author>
 }
